refactor: improve nmrfile-browser
